chore(styles): separate all styles to other files
